+++ CHANGELOG.rst
@@ -7,13 +7,17 @@ Changelog
    but in reStructuredText instead of Markdown (for ease of incorporation into
    Sphinx documentation and the PyPI description).
 
-2025-04-25
+2025-04-28
 **********
 Added
 =====
 * Send a slug with API test calls to help distinguish them from normal traffic.
 
-2025-04-18
+Fixed
+=====
+* Make API tests compatible with Python 3.11 and 3.12
+
+2025-04-22
 **********
 Fixed
 =====